requirements.txt
discord.py>=2.3,<3.0
fastapi>=0.104,<1.0
uvicorn[standard]>=0.24,<1.0
aiosqlite>=0.19,<1.0
chromadb>=0.4,<1.0
apscheduler>=3.10,<4.0
httpx>=0.25,<1.0
pyyaml>=6.0,<7.0
google-generativeai>=0.8,<1.0
